development | cambios capitulo 23 y 24 autowired

+++ src/main/java/com/platzimarket/domain/repository/IProductRepository.java
@@ -1,6 +1,7 @@
 package com.platzimarket.domain.repository;
 
 import com.platzimarket.domain.Product;
+import com.platzimarket.persistence.entity.ProductEntity;
 
 import java.util.List;
 import java.util.Optional;
@@ -13,9 +14,9 @@ public interface IProductRepository {
 
   Optional<List<Product>> getScarseProducts(int quantity);
 
-  Optional<Product> getProduct(int productId);
+  Optional<Product> getProduct(long productId);
 
   Product save(Product product);
 
-  void delete(int productId);
+  void delete(long productId);
 }
